requirements.txt
altgraph==0.17.2
asgiref==3.5.2
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
pefile==2022.5.30
psycopg2==2.9.3
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
python-decouple==3.6
pytz==2022.1
pywin32-ctypes==0.2.0
sqlparse==0.4.2
tzdata==2022.1
unicode==2.9
Unidecode==1.3.4
